+++ studyParty_group/src/main/java/com/studyparty/group/mapper/GroupMapper.java
@@ -8,5 +8,5 @@ import java.util.List;
 
 @Mapper
 public interface GroupMapper extends BaseMapper<Group> {
-    List<Group> findMyGroups(int id);
+    List<Group> findMyGroups(Long id);
 }
