+++ studyParty_group/src/main/java/com/studyparty/group/services/GroupJoinServer.java
@@ -7,8 +7,8 @@ import com.studyParty.entity.group.GroupJoin;
 import java.util.List;
 
 public interface GroupJoinServer extends IService<GroupJoin> {
-    List<List<GroupJoin>> findMyGroups(int userId);
-    boolean agreeJoin(int id, int userId, GroupJoin groupJoin);
-    boolean disagreeJoin(int id, int userId, GroupJoin groupJoin);
-    boolean isJoined(int groupId, int userId);
+    List<List<GroupJoin>> findMyGroups(Long userId);
+    boolean agreeJoin(Long id, Long userId, GroupJoin groupJoin);
+    boolean disagreeJoin(Long id, Long userId, GroupJoin groupJoin);
+    boolean isJoined(Long groupId, Long userId);
 }
