+++ studyParty_group/src/main/java/com/studyparty/group/services/GroupServer.java
@@ -7,5 +7,5 @@ import com.studyParty.entity.group.Group;
 import java.util.List;
 
 public interface GroupServer extends IService<Group> {
-    List<Group> findMyGroups(int id);
+    List<Group> findMyGroups(Long id);
 }
